+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.1
+
+- Empathy Cores damage boost only affects skill damage instead of base damage.
+- War Bonds is no longer inheritable.
+- Eccentric Vase
+    - Cooldown decreased. (45s -> 30s)
+    - Distance increased. (1000m -> 2000m)
+    - Improved acceleration while traveling.
+    - Grants invisibility & intangibility while traveling.
+    - Grants brief (0.5s) fall damage immunity on exit.
+- Molotov (6-Pack)
+    - Fire residue duration increased. (7s -> 10s)
+    - Fire residue size multiplied by x1.75.
+    - Fire residue tickrate increased to 2/s. Believe it or not, it's actually just 1/s right now.
+    - Added a config to adjust the impact explosion damage, though it is unchanged by default.
+
 ## 1.0.0
 
 - Release.